Support `WATCH_NAMESPACE` in List and Watch requests
As it often gets, writing new feature is easier than writing its tests.
First of all I had to put `WATCH_NAMESPACE` tests into a separate package
because they required a different operator configuration. The e2e test
setup code was moved into a separate auxiliary package in order to reuse
it. There was a warning in Ginkgo that `func(Done)` actions are
deprecated so I've fixed it.