Add `WATCH_NAMESPACE` environment variable support (#329)
* Add `WATCH_NAMESPACE` environment variable support

It is sometimes required to contain operator activity inside a fixed set
of namespaces. There is a documented common practice in `operator-sdk`
to use `WATCH_NAMESPACE` environment variable for that purpose.
Furthermore, `WATCH_NAMESPACE` is already used in kustomization and
Makefile but is being silently ignored.

* Support `WATCH_NAMESPACE` in List and Watch requests

As it often gets, writing new feature is easier than writing its tests.
First of all I had to put `WATCH_NAMESPACE` tests into a separate package
because they required a different operator configuration. The e2e test
setup code was moved into a separate auxiliary package in order to reuse
it. There was a warning in Ginkgo that `func(Done)` actions are
deprecated so I've fixed it.