controllers: properly perform objects delete operation
Previously, it was not possible to detect should additional Object must be deleted.
Such as `HPA` or `PodDisruptionBudget`, because operator didn't have enough context for that.
If current object spec is nil, it was not possible to check it's state before changes were applied.

 This commit, adds a check for previous Spec state and performs delete operations based on it.
It allows to keep unmanaged by operator resources with the same names. And it should reduce amount of DELETE api requests.

 Related issues:
https://github.com/VictoriaMetrics/operator/issues/758
https://github.com/VictoriaMetrics/operator/issues/817

Signed-off-by: f41gh7 <nik@victoriametrics.com>