operator: rework structured logging
Current logger lig `logr` doesn't support formatted message field. Because of that
we had to use structured logging with key-value pairs. It has performance benefits.
But it also produces additional load into the logging systems. And the most importantly
it creats additional congnitive load on users. Since it's hard to join json fields
to get the context of the log message.

 This commit adds the following changes:
* use fmt.Sprintf in order to add more context to the message body.
* remove irrelevant fields from WithValues, like `secret_for` field
* change logger name to `controller.CRD_NAME` format.
* change child objects parent_NAME: OBJECT_NAMe to simple NAME: OBJECT_NAME. It allows to easily query
 log messages by pattern.

Related issue:
https://github.com/VictoriaMetrics/operator/issues/1191

Signed-off-by: f41gh7 <nik@victoriametrics.com>