build/vmcluster: properly mount ServiceAccount for `vmauth` load-balancer
Initially `ServiceAccount` wasn't defined for `vmauth`'s `Deployment`. Since there was no practial use for it.
At older versions of operator it made sense to mount `serviceAccount` for `vminsert` and `vmselect` components because of `PodSecurityPolicy`
Kubernetes resources. But it's deprecated now and it makes `serviceAccount` irrelevant for these components.

 But for production use cases it could be useful. Strict security policy may require to disable auto service account mount.
On of possible solutions for it to use `ServiceAccount` with disabled auto-mount option.

 This commit adds `serviceAccount` of `VMCluster` to the load-balancer `Deployment`.

 Related issue:
https://github.com/VictoriaMetrics/operator/issues/1210

Signed-off-by: f41gh7 <nik@victoriametrics.com>