operator: adds new flag controller.statusLastUpdateTimeTTL
It controlls expiration time of status.condition lastUpdateTime. Which is needed to track stale parent objects.
Increasing value of this flag reduces load on Kubernetes cluster, but it also increases time of stale object detection.

 For instance, if there are 2 VMAlert objects and it matches some VMRule. Both vmalerts will be registered at VMRule.status.conditions[].type with it's name.
In case when 1 of VMAlert objects were deleted, it will be removed from VMRule.status.condition only after 3*controller.statusLastUpdateTimeTTL. Which take up to 3 hours
with default values.

 Related issue:
https://github.com/VictoriaMetrics/operator/issues/1220

Signed-off-by: f41gh7 <nik@victoriametrics.com>