docs: Enchance requests load balancing documentation
This enhances the clarity around how the requestsLoadBalancer option
works by:

Adding cross-references to the detailed documentation on request
load balancing and the corresponding API spec option.
Introducing a new paragraph that highlights scenarios where vmauth
could be used as an alternative to requestsLoadBalancer.

This is a follow-up to issue #2024.