cmd/config-reloader: properly configure web-server
 Previously, if `reload-use-proxy-protocol` is set for config-reloader,
it started to listen built-in web-server with proxy-protocol. It's not
intentional and it makes impossible to use health probes for it.
The purpose of reload-use-proxy-protocol flag, is to configure
http-client to use proxy-protocol for main app API requests.

 This commit removes this flag from web-server and changes it from Bool
array into boolean flag. Allows to use proxy-protocol at main app.

Related issue:
https://github.com/VictoriaMetrics/operator/issues/1308