factory/vmanomaly/config: fix marshalling of writer.metricFormat
Previously, it was not possible to add custom labels for writer. Replaced struct to a map of values since it was not possible to embed both named fields and a map, that would cause a failure of api-gen.

Implemented migration of previously supported `name` to `__name__` field as vmanomaly expects a specific name.

See https://docs.victoriametrics.com/anomaly-detection/components/writer/#metrics-formatting for reference of vmanomaly config.