api/vmcluster: introduce new field rollingUpdateStrategyBehavior
This commit introduces a new VMCluster CRD property called maxUnavailable in:

    spec.vmstorage.rollingUpdateStrategyBehavior.maxUnavailable
    spec.vmselect.rollingUpdateStrategyBehavior.maxUnavailable

It's very similar to this (alpha) feature of Kubernetes:
https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#maximum-unavailable-pods

This commit makes it work with the OnDelete upgrade policy used in the operator by default.

Additionally, this PR makes it available for much older Kubernetes versions or, for example, it makes it finally available for those running GKE clusters that don't support alpha features.

Some examples for the maxUnavailable field:

    maxUnavailable: 1 - default, which mostly (see details on PDBs below) follows the current behavior of the operator.
    maxUnavailable: 2 - for cases when -replicationFactor=3.
    maxUnavailable: 100% - for cases when the "minimum downtime strategy" is preferable.

Since more than one pod can now become unavailable during an upgrade (i.e. voluntary disruptions) it's important to respect a PDB configuration to some degree. So this PR introduces a partial support for PDBs targeted at StatefulSets of VMCluster according to this:
https://kubernetes.io/docs/tasks/run-application/configure-pdb/#arbitrary-controllers-and-selectors

This is achieved with the Eviction API instead of directly deleting Pods, as described here:
https://kubernetes.io/docs/concepts/scheduling-eviction/api-eviction/

And since it appears that "write" RBAC permission for Pods is no longer needed, I've moved it to the "read-only" section.

Fixes https://github.com/VictoriaMetrics/operator/issues/1457